move PhaseEstimationCircuit to algos
